+++ display_list/BUILD.gn
@@ -63,6 +63,8 @@ source_set("display_list") {
     "skia/dl_sk_canvas.h",
     "skia/dl_sk_dispatcher.cc",
     "skia/dl_sk_dispatcher.h",
+    "skia/dl_sk_utils.cc",
+    "skia/dl_sk_utils.h",
     "types.h",
   ]
 
@@ -102,8 +104,8 @@ if (enable_unittests) {
       "display_list_path_effect_unittests.cc",
       "display_list_rtree_unittests.cc",
       "display_list_unittests.cc",
-      "display_list_utils_unittests.cc",
       "display_list_vertices_unittests.cc",
+      "skia/dl_sk_utils_unittests.cc",
     ]
 
     deps = [
